+++ crates/ryo3-jiff/src/lib.rs
@@ -21,6 +21,8 @@ mod deprecations;
 mod dev;
 mod difference;
 mod errors;
+mod from;
+mod from_str;
 mod functions;
 mod interns;
 mod into_span_arithmetic;