+++ crates/ryo3-jiff/src/ry_date.rs
@@ -1,4 +1,3 @@
-use crate::constants::DATETIME_PARSER;
 use crate::difference::{DateDifferenceArg, RyDateDifference};
 use crate::errors::{map_py_overflow_err, map_py_value_err};
 use crate::ry_datetime::RyDateTime;
@@ -64,15 +63,14 @@ impl RyDate {
 
     #[staticmethod]
     fn from_str(s: &str) -> PyResult<Self> {
-        DATETIME_PARSER
-            .parse_date(s)
-            .map(Self::from)
-            .map_err(map_py_value_err)
+        use ryo3_core::PyFromStr;
+        Self::py_from_str(s)
     }
 
     #[staticmethod]
-    fn parse(s: &str) -> PyResult<Self> {
-        Self::from_str(s)
+    fn parse(s: &Bound<'_, PyAny>) -> PyResult<Self> {
+        use ryo3_core::PyParse;
+        Self::py_parse(s)
     }
 
     #[pyo3(signature = (hour, minute, second, nanosecond=0))]
@@ -528,11 +526,5 @@ impl Display for RyDate {
             self.month(),
             self.day()
         )
-    }
-}
-
-impl From<Date> for RyDate {
-    fn from(value: Date) -> Self {
-        Self(value)
     }
 }