+++ crates/ryo3-jiff/src/ry_span.rs
@@ -11,7 +11,6 @@ use pyo3::types::{PyDelta, PyDict, PyFloat, PyInt, PyTuple};
 use ryo3_macro_rules::{any_repr, py_overflow_error, py_type_err, py_value_error};
 use std::fmt::Display;
 use std::hash::{DefaultHasher, Hash, Hasher};
-use std::str::FromStr;
 
 #[cfg_attr(feature = "serde", derive(serde::Serialize))]
 #[cfg_attr(feature = "serde", serde(transparent))]
@@ -187,12 +186,14 @@ impl RySpan {
 
     #[staticmethod]
     fn from_str(s: &str) -> PyResult<Self> {
-        Span::from_str(s).map(Self::from).map_err(map_py_value_err)
+        use ryo3_core::PyFromStr;
+        Self::py_from_str(s)
     }
 
     #[staticmethod]
-    fn parse(s: &str) -> PyResult<Self> {
-        Self::from_str(s)
+    fn parse(s: &Bound<'_, PyAny>) -> PyResult<Self> {
+        use ryo3_core::PyParse;
+        Self::py_parse(s)
     }
 
     #[expect(clippy::too_many_arguments)]
@@ -825,28 +826,6 @@ impl Display for RySpan {
         }
 
         f.write_str(")")
-    }
-}
-
-impl From<Span> for RySpan {
-    fn from(span: Span) -> Self {
-        Self(span)
-    }
-}
-
-impl From<JiffSpan> for RySpan {
-    fn from(span: JiffSpan) -> Self {
-        Self(span.0)
-    }
-}
-
-impl TryFrom<SignedDuration> for RySpan {
-    type Error = PyErr;
-
-    fn try_from(value: SignedDuration) -> Result<Self, Self::Error> {
-        Span::try_from(value)
-            .map(Self::from)
-            .map_err(map_py_overflow_err)
     }
 }