+++ crates/ryo3-jiff/src/ry_timestamp.rs
@@ -18,7 +18,6 @@ use ryo3_macro_rules::{any_repr, py_type_error};
 use std::fmt::Display;
 use std::hash::{DefaultHasher, Hash, Hasher};
 use std::ops::Sub;
-use std::str::FromStr;
 
 #[cfg_attr(feature = "serde", derive(serde::Serialize))]
 #[cfg_attr(feature = "serde", serde(transparent))]
@@ -74,14 +73,14 @@ impl RyTimestamp {
 
     #[staticmethod]
     fn from_str(s: &str) -> PyResult<Self> {
-        Timestamp::from_str(s)
-            .map(Self::from)
-            .map_err(map_py_value_err)
+        use ryo3_core::PyFromStr;
+        Self::py_from_str(s)
     }
 
     #[staticmethod]
-    fn parse(s: &str) -> PyResult<Self> {
-        Self::from_str(s)
+    fn parse(s: &Bound<'_, PyAny>) -> PyResult<Self> {
+        use ryo3_core::PyParse;
+        Self::py_parse(s)
     }
 
     #[staticmethod]
@@ -471,11 +470,5 @@ impl Display for RyTimestamp {
             self.0.as_second(),
             self.0.subsec_nanosecond()
         )
-    }
-}
-
-impl From<Timestamp> for RyTimestamp {
-    fn from(value: Timestamp) -> Self {
-        Self(value)
     }
 }